python-api/requirements.txt
torch>=2.2.0
torchvision>=0.17.0
facenet-pytorch==2.5.3
efficientnet-pytorch==0.7.1
tensorflow>=2.15.0
librosa==0.10.1
fastapi==0.104.1
python-multipart==0.0.6
uvicorn==0.24.0
numpy>=1.24.3
Pillow>=10.1.0
opencv-python>=4.8.1.78
faker==20.1.0
